Changed DrawOnPictureBox in ImageDraw to use Graphics object instead of my own nested for loop code for drawing. This way we can eventually change the shape of the line (i.e. rounded, square, etc.). Added Draw button to GUI that toggles drawing capability on and off (set to disabled initially).
Found bug with ImageResize where image was not being set, fixed.

Beware: sometimes the ShoppedGui.Designer.cs likes to delete the MouseUp event on PictureBox, which will cause our code to not work properly. If Undo/Redo seem to not be working correctly with Draw, check this file to make sure the event is still there.